Misc. Updates
Changed some locations TextField to ForeignKeyField so that they reference an actual location in the database. Didn't do that for Calendars, as the location in the calendar might not correspond, or have GPS coordinates, so left it as TextField for now.

Update the YAML file to include the fields that were missing

Add boolean field for a message including media, better see what is included. Not currently planning on making a table for media, storing the files in the SQLite would take too much space.